[sample/SampleRobot] use ''  to specify model location
